added unit test for when there is only one dependency
