+++ src/index.js
@@ -3,7 +3,7 @@ export { ErrorContext } from './models/errorContext';
 
 // Middleware
 export { errorHandler } from './middleware/errorHandler';
-export { healthRoutes } from './middleware/health';
+export { healthRoutes } from './middleware/healthRoutes';
 
 // Test Utils
 export { FakeAuthProvider } from './test-utils/fakeAuthProvider';
